+++ web/src/pages/Calendar.tsx
@@ -1,9 +1,0 @@
-import { InvoiceCalendar } from '../components/InvoiceCalendar';
-
-export function Calendar() {
-  return (
-    <div className="space-y-6">
-      <InvoiceCalendar />
-    </div>
-  );
-}